Split address-specific code into their own files
Since different environemnts support different types of sockets, this
makes it easier to modularize the code in the future (e.g., if I ever
decide to try and support Windows and need to remote unix sockets).  As
a result, a couple files (domain.hpp and sockaddr.hpp) don't serve any
purpose at this point, and have been removed.